Refactor
Use refs for avoiding excess message display.
